requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
idna==3.2
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
threadpoolctl==2.2.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.10.0.0
urllib3==1.26.6
